Added CMake support to build project.
There is one change from the original Makefile. By default project
was installed to $PREFIX/sbin and now its $PREFIX/bin.

"sbin" directory is meant to contain boot executables, or
executables intended to run from root user. As we dont want to run
this software from root so the change was made.

Also some distributions does not contain "sbin" in PATH for normal
users and this change is handy.